test(zellij): move the writer fixture beside its presence callers
