+++ crates/rimz/src/mux/zellij/tests/support.rs
@@ -11,8 +11,6 @@ use tempfile::TempDir;
 
 use crate::ids::WorkspaceId;
 use crate::mux::PresencePluginOptions;
-use crate::mux::zellij::pane_topology::TopologyWriter;
-use crate::mux::zellij::{presence_plugin_build, presence_plugin_config_hash_for};
 
 /// Write `script` as an executable `zellij` beside a fresh temp dir.
 #[cfg(unix)]
@@ -85,17 +83,5 @@ pub(crate) fn presence_opts(session_name: &str, rimz_bin: &str) -> PresencePlugi
         zoom_key: None,
         focus_follows_mouse: false,
         mouse_click_through: true,
-    }
-}
-
-/// A writer record carrying this host's build and config identity — what the
-/// presence retire path accepts as proof that a replacement plugin is live.
-pub(crate) fn current_writer(plugin_id: u32, loaded_at_ms: u64) -> TopologyWriter {
-    let opts = presence_opts("rimz-test", "/home/user/.cargo/bin/rimz");
-    TopologyWriter {
-        plugin_id,
-        loaded_at_ms,
-        build: Some(presence_plugin_build().to_owned()),
-        config: Some(presence_plugin_config_hash_for(&opts)),
     }
 }
